Make an attempt at organizing hdwallet provider files
