+++ packages/hdwallet-provider/src/index.ts
@@ -18,6 +18,9 @@ import WebsocketProvider from "@trufflesuite/web3-provider-engine/subproviders/w
 import Url from "url";
 import { JSONRPCRequestPayload, JSONRPCErrorCallback } from "ethereum-protocol";
 import { Callback, JsonRPCResponse } from "web3/providers";
+import { ConstructorArguments } from "./types";
+import { NewConstructor } from "./NewConstructor";
+import { getOptions } from "./getOptions";
 
 // Important: do not use debug module. Reason: https://github.com/trufflesuite/truffle/issues/2374#issuecomment-536109086
 
@@ -27,228 +30,6 @@ import { Callback, JsonRPCResponse } from "web3/providers";
 // of this behavior by passing `shareNonce=false` to the constructor.
 // See issue #65 for more
 const singletonNonceSubProvider = new NonceSubProvider();
-
-/*
- * type aliases for better readability around legacy positional arguments
- */
-export type MnemonicPhrase = string;
-export type MnemonicPassword = string;
-export interface Mnemonic {
-  phrase: MnemonicPhrase;
-  password?: MnemonicPassword;
-}
-export type PrivateKey = string;
-export type Provider = any;
-export type ProviderUrl = string;
-export type ProviderOrUrl = Provider | ProviderUrl;
-export type AddressIndex = number;
-export type NumberOfAddresses = number;
-export type ShareNonce = boolean;
-export type DerivationPath = string;
-
-/*
- * namespace wrapper for new constructor options interface
- */
-export namespace NewConstructor {
-  export interface MnemonicSigningAuthority {
-    mnemonic: Mnemonic;
-  }
-
-  export interface PrivateKeySigningAuthority {
-    privateKeys: PrivateKey[];
-  }
-
-  export type SigningAuthority =
-    | MnemonicSigningAuthority
-    | PrivateKeySigningAuthority;
-
-  export interface CommonOptions {
-    providerOrUrl: ProviderOrUrl;
-    addressIndex?: AddressIndex;
-    numberOfAddresses?: NumberOfAddresses;
-    shareNonce?: ShareNonce;
-    derivationPath?: DerivationPath;
-  }
-
-  export type Options = SigningAuthority & CommonOptions;
-
-  // extract the mnemonic if that's the style used, or return undefined
-  export const getMnemonic = (
-    signingAuthority: SigningAuthority
-  ): Mnemonic | undefined => {
-    if ("mnemonic" in signingAuthority) {
-      return signingAuthority.mnemonic;
-    }
-  };
-
-  // extract the private keys if that's the style used, or return undefined
-  export const getPrivateKeys = (
-    signingAuthority: SigningAuthority
-  ): PrivateKey[] | undefined => {
-    if ("privateKeys" in signingAuthority) {
-      return signingAuthority.privateKeys;
-    }
-  };
-}
-
-/*
- * namespace wrapper for old-style positional arguments
- */
-export namespace LegacyConstructor {
-  type PossibleArguments = [
-    /*
-     * required
-     */
-    MnemonicPhrase | PrivateKey[],
-    ProviderOrUrl,
-
-    /*
-     * optional
-     */
-    AddressIndex,
-    NumberOfAddresses,
-    ShareNonce,
-    DerivationPath
-  ];
-
-  // (awful to have to do it this way)
-  export type Arguments =
-    | [PossibleArguments[0], PossibleArguments[1]]
-    | [PossibleArguments[0], PossibleArguments[1], PossibleArguments[2]]
-    | [
-        PossibleArguments[0],
-        PossibleArguments[1],
-        PossibleArguments[2],
-        PossibleArguments[3]
-      ]
-    | [
-        PossibleArguments[0],
-        PossibleArguments[1],
-        PossibleArguments[2],
-        PossibleArguments[3],
-        PossibleArguments[4]
-      ]
-    | [
-        PossibleArguments[0],
-        PossibleArguments[1],
-        PossibleArguments[2],
-        PossibleArguments[3],
-        PossibleArguments[4],
-        PossibleArguments[5]
-      ];
-
-  // check that the first argument is a mnemonic phrase
-  const isMnemonicPhrase = (
-    mnemonicPhraseOrPrivateKeys: MnemonicPhrase | PrivateKey[]
-  ): mnemonicPhraseOrPrivateKeys is MnemonicPhrase =>
-    typeof mnemonicPhraseOrPrivateKeys === "string";
-
-  // check that the first argument is a list of private keys
-  const isPrivateKeys = (
-    mnemonicPhraseOrPrivateKeys: MnemonicPhrase | PrivateKey[]
-  ): mnemonicPhraseOrPrivateKeys is PrivateKey[] =>
-    mnemonicPhraseOrPrivateKeys instanceof Array;
-
-  // turn polymorphic first argument into { mnemonic } or { privateKeys }
-  const getSigningAuthorityOptions = (
-    mnemonicPhraseOrPrivateKeys: MnemonicPhrase | PrivateKey[]
-  ): NewConstructor.SigningAuthority => {
-    if (isMnemonicPhrase(mnemonicPhraseOrPrivateKeys)) {
-      return {
-        mnemonic: {
-          phrase: mnemonicPhraseOrPrivateKeys
-        }
-      };
-    } else if (isPrivateKeys(mnemonicPhraseOrPrivateKeys)) {
-      return {
-        privateKeys: mnemonicPhraseOrPrivateKeys
-      };
-    } else {
-      throw new Error(
-        `First argument to new HDWalletProvider() must be a mnemonic phrase or a list of private keys. ` +
-          `Received: ${JSON.stringify(mnemonicPhraseOrPrivateKeys)}`
-      );
-    }
-  };
-
-  // convert legacy style positional arguments to new, single-arg options format
-  export const toOptions = (args: Arguments): NewConstructor.Options => {
-    // otherwise, if arguments match the old-style, extract properties and handle polymorphism
-    const [
-      mnemonicPhraseOrPrivateKeys,
-      providerOrUrl,
-      addressIndex,
-      numberOfAddresses,
-      shareNonce,
-      derivationPath
-    ] = args;
-
-    const signingAuthority = getSigningAuthorityOptions(
-      mnemonicPhraseOrPrivateKeys
-    );
-
-    return {
-      ...signingAuthority,
-      providerOrUrl,
-      addressIndex,
-      numberOfAddresses,
-      shareNonce,
-      derivationPath
-    };
-  };
-}
-
-/*
- * top-level polymorphic type
- */
-export type ConstructorArguments =
-  | LegacyConstructor.Arguments // either the old-style tuple type
-  | [NewConstructor.Options]; // or a single argument for new-style options
-
-// type predicate guard to determine at runtime if arguments conform to
-// new-style constructor args.
-const matchesNewOptions = (
-  args: ConstructorArguments
-): args is [NewConstructor.Options] => {
-  // new-style means exactly one argument
-  if (args.length !== 1) {
-    return false;
-  }
-
-  const [options] = args;
-
-  // beyond that, determine based on property inclusion check for required keys
-  return (
-    "providerOrUrl" in options &&
-    ("mnemonic" in options || "privateKeys" in options)
-  );
-};
-
-// type predicate guard to determine at runtime if arguments conform to
-// old-style constructor args.
-const matchesLegacyArguments = (
-  args: ConstructorArguments
-): args is LegacyConstructor.Arguments =>
-  // first check for alternate (new-style) case for basic determination
-  !matchesNewOptions(args) &&
-  // then additionally make sure we have the two required options we need
-  args.filter(arg => arg !== undefined).length >= 2;
-
-// normalize arguments passed to constructor to match single, new-style options
-// argument
-const getOptions = (...args: ConstructorArguments): NewConstructor.Options => {
-  if (matchesNewOptions(args)) {
-    // if arguments already match new-style, no real transformation needed
-    const [options] = args;
-    return options;
-  } else if (matchesLegacyArguments(args)) {
-    return LegacyConstructor.toOptions(args);
-  } else {
-    throw new Error(
-      "Unknown arguments format passed to new HDWalletProvider. Please check your configuration and try again"
-    );
-  }
-};
 
 class HDWalletProvider {
   private hdwallet?: EthereumHDKey;
